backend/requirements.txt
amqp==2.6.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==4.4.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
drf-yasg==1.21.4
flake8==6.0.0
flower==0.9.7
gevent==22.10.2
greenlet==2.0.1
humanize==4.4.0
idna==3.4
inflection==0.5.1
isort==5.11.1
itypes==1.2.0
Jinja2==3.1.2
kombu==4.6.11
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==22.0
phonenumbers==8.13.2
Pillow==9.3.0
prometheus-client==0.8.0
prompt-toolkit==3.0.36
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2022.6
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==1.3.0
wcwidth==0.2.5
zope.event==4.6
zope.interface==5.5.2
